requirements.txt
cycler==0.11.0
fonttools==4.28.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
mysql-connector-python==8.0.27
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tomli==1.2.2